+++ app/upgrades/v1/mainnet/allowed_staking_tx_hashes.go
@@ -1,0 +1,9 @@
+package mainnet
+
+// TODO: Specify allowed staking tx hashes for v1 upgrade for mainnet
+const AllowedStakingTxHashesStr = `{
+	"tx_hashes": [
+		"90463a1e0b19137f07c304f2c1c79351fa6f97f1195f1f90dc0b3b57893a350d",
+		"1db6de324beefd6870ec62a7385cac7144a43eba8e61fca202a8b41d963680ef"
+	]
+}`