+++ app/upgrades/v1/upgrades_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/babylonlabs-io/babylon/test/e2e/util"
 	"github.com/babylonlabs-io/babylon/testutil/datagen"
 	minttypes "github.com/babylonlabs-io/babylon/x/mint/types"
+	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -45,18 +46,22 @@ var (
 	wasmContract []byte
 
 	UpgradeV1DataTestnet = v1.UpgradeDataString{
-		BtcStakingParamStr:    testnetdata.BtcStakingParamStr,
-		FinalityParamStr:      testnetdata.FinalityParamStr,
-		CosmWasmParamStr:      testnetdata.CosmWasmParamStr,
-		NewBtcHeadersStr:      testnetdata.NewBtcHeadersStr,
-		TokensDistributionStr: testnetdata.TokensDistributionStr,
+		BtcStakingParamStr:        testnetdata.BtcStakingParamStr,
+		FinalityParamStr:          testnetdata.FinalityParamStr,
+		IncentiveParamStr:         testnetdata.IncentiveParamStr,
+		CosmWasmParamStr:          testnetdata.CosmWasmParamStr,
+		NewBtcHeadersStr:          testnetdata.NewBtcHeadersStr,
+		TokensDistributionStr:     testnetdata.TokensDistributionStr,
+		AllowedStakingTxHashesStr: testnetdata.AllowedStakingTxHashesStr,
 	}
 	UpgradeV1DataMainnet = v1.UpgradeDataString{
-		BtcStakingParamStr:    mainnetdata.BtcStakingParamStr,
-		FinalityParamStr:      mainnetdata.FinalityParamStr,
-		CosmWasmParamStr:      mainnetdata.CosmWasmParamStr,
-		NewBtcHeadersStr:      mainnetdata.NewBtcHeadersStr,
-		TokensDistributionStr: mainnetdata.TokensDistributionStr,
+		BtcStakingParamStr:        mainnetdata.BtcStakingParamStr,
+		FinalityParamStr:          mainnetdata.FinalityParamStr,
+		IncentiveParamStr:         mainnetdata.IncentiveParamStr,
+		CosmWasmParamStr:          mainnetdata.CosmWasmParamStr,
+		NewBtcHeadersStr:          mainnetdata.NewBtcHeadersStr,
+		TokensDistributionStr:     mainnetdata.TokensDistributionStr,
+		AllowedStakingTxHashesStr: mainnetdata.AllowedStakingTxHashesStr,
 	}
 	UpgradeV1Data = []v1.UpgradeDataString{UpgradeV1DataTestnet, UpgradeV1DataMainnet}
 )
@@ -258,6 +263,12 @@ func (s *UpgradeTestSuite) PostUpgrade() {
 	newHeadersLen := len(allBtcHeaders)
 	s.Equal(newHeadersLen, s.btcHeadersLenPreUpgrade+lenHeadersInserted)
 
+	// ensure the incentive params were set as expected
+	incentiveParamsFromUpgrade, err := v1.LoadIncentiveParamsFromData(s.app.AppCodec(), s.upgradeDataStr.IncentiveParamStr)
+	s.NoError(err)
+	incentiveParams := s.app.IncentiveKeeper.GetParams(s.ctx)
+	s.EqualValues(incentiveParamsFromUpgrade, incentiveParams)
+
 	// ensure the headers were inserted as expected
 	for i, btcHeaderInserted := range btcHeadersInserted {
 		btcHeaderInState := allBtcHeaders[s.btcHeadersLenPreUpgrade+i]
@@ -292,4 +303,19 @@ func (s *UpgradeTestSuite) PostUpgrade() {
 	upgradeWasmParams, err := v1.LoadCosmWasmParamsFromData(s.app.AppCodec(), s.upgradeDataStr.CosmWasmParamStr)
 	s.NoError(err)
 	s.EqualValues(chainWasmParams, upgradeWasmParams)
+
+	allowedStakingTxHashes, err := v1.LoadAllowedStakingTransactionHashesFromData(s.upgradeDataStr.AllowedStakingTxHashesStr)
+	s.NoError(err)
+	s.NotNil(allowedStakingTxHashes)
+	s.Greater(len(allowedStakingTxHashes.TxHashes), 0)
+
+	for _, txHash := range allowedStakingTxHashes.TxHashes {
+		hash, err := chainhash.NewHashFromStr(txHash)
+		s.NoError(err)
+
+		s.True(s.app.BTCStakingKeeper.IsStakingTransactionAllowed(s.ctx, hash))
+	}
+
+	nonExistentTxHash := chainhash.Hash{}
+	s.False(s.app.BTCStakingKeeper.IsStakingTransactionAllowed(s.ctx, &nonExistentTxHash))
 }