+++ x/btcstaking/keeper/btc_delegations.go
@@ -53,7 +53,6 @@ func (k Keeper) AddBTCDelegation(
 	k.setBTCDelegation(ctx, btcDel)
 
 	if err := ctx.EventManager().EmitTypedEvents(types.NewBtcDelCreationEvent(
-		stakingTxHash.String(),
 		btcDel,
 	)); err != nil {
 		panic(fmt.Errorf("failed to emit events for the new pending BTC delegation: %w", err))
@@ -80,7 +79,7 @@ func (k Keeper) AddBTCDelegation(
 			NewState:      types.BTCDelegationStatus_UNBONDED,
 		})
 
-		// NOTE: we should have verified that EndHeight > btcTip.Height + max(w, min_unbonding_time)
+		// NOTE: we should have verified that EndHeight > btcTip.Height + min_unbonding_time
 		k.addPowerDistUpdateEvent(ctx, btcDel.EndHeight-minUnbondingTime, unbondedEvent)
 	}