+++ x/finality/keeper/power_dist_change_test.go
@@ -62,6 +62,7 @@ func FuzzProcessAllPowerDistUpdateEvents_Determinism(f *testing.F) {
 					0,
 					0,
 					false,
+					false,
 				)
 				h.NoError(err)
 				event := types.NewEventPowerDistUpdateWithBTCDel(&types.EventBTCDelegationStateUpdate{
@@ -119,6 +120,7 @@ func FuzzSlashFinalityProviderEvent(f *testing.F) {
 			0,
 			0,
 			true,
+			false,
 		)
 		h.NoError(err)
 		// give it a quorum number of covenant signatures
@@ -205,6 +207,7 @@ func FuzzJailFinalityProviderEvents(f *testing.F) {
 			0,
 			0,
 			true,
+			false,
 		)
 		h.NoError(err)
 		// give it a quorum number of covenant signatures
@@ -274,6 +277,7 @@ func FuzzJailFinalityProviderEvents(f *testing.F) {
 			0,
 			0,
 			true,
+			false,
 		)
 		h.NoError(err)
 		// give it a quorum number of covenant signatures
@@ -335,6 +339,7 @@ func FuzzUnjailFinalityProviderEvents(f *testing.F) {
 			0,
 			0,
 			true,
+			false,
 		)
 		h.NoError(err)
 		// give it a quorum number of covenant signatures
@@ -426,6 +431,8 @@ func FuzzBTCDelegationEvents_NoPreApproval(f *testing.F) {
 		stakingValue := int64(2 * 10e8)
 		delSK, _, err := datagen.GenRandomBTCKeyPair(r)
 		h.NoError(err)
+
+		stakingParams := h.BTCStakingKeeper.GetParamsWithVersion(h.Ctx).Params
 		stakingTxHash, msgCreateBTCDel, actualDel, _, _, _, err := h.CreateDelegation(
 			r,
 			delSK,
@@ -435,6 +442,7 @@ func FuzzBTCDelegationEvents_NoPreApproval(f *testing.F) {
 			1000,
 			0,
 			0,
+			false,
 			false,
 		)
 		h.NoError(err)
@@ -448,13 +456,8 @@ func FuzzBTCDelegationEvents_NoPreApproval(f *testing.F) {
 		events := h.BTCStakingKeeper.GetAllPowerDistUpdateEvents(h.Ctx, btcTip.Height, btcTip.Height)
 		require.Len(t, events, 0)
 
-		btckptParams := btccKeeper.GetParams(h.Ctx)
-		stakingParams := h.BTCStakingKeeper.GetParamsWithVersion(h.Ctx).Params
-
-		minUnbondingTime := types.MinimumUnbondingTime(&stakingParams, &btckptParams)
-
-		// the BTC delegation will be unbonded at end height - max(w, min_unbonding_time)
-		unbondedHeight := actualDel.EndHeight - minUnbondingTime
+		// the BTC delegation will be unbonded at end height - min_unbonding_time
+		unbondedHeight := actualDel.EndHeight - stakingParams.MinUnbondingTimeBlocks
 		events = h.BTCStakingKeeper.GetAllPowerDistUpdateEvents(h.Ctx, unbondedHeight, unbondedHeight)
 		require.Len(t, events, 1)
 		btcDelStateUpdate := events[0].GetBtcDelStateUpdate()
@@ -554,6 +557,7 @@ func FuzzBTCDelegationEvents_WithPreApproval(f *testing.F) {
 			0,
 			0,
 			true,
+			false,
 		)
 		h.NoError(err)
 
@@ -599,13 +603,8 @@ func FuzzBTCDelegationEvents_WithPreApproval(f *testing.F) {
 		require.Equal(t, stakingTxHash, btcDelStateUpdate.StakingTxHash)
 		require.Equal(t, types.BTCDelegationStatus_ACTIVE, btcDelStateUpdate.NewState)
 
-		btckptParams := btccKeeper.GetParams(h.Ctx)
-		stakingParams := h.BTCStakingKeeper.GetParamsWithVersion(h.Ctx).Params
-
-		minUnbondingTime := types.MinimumUnbondingTime(&stakingParams, &btckptParams)
-
-		// the BTC delegation will be unbonded at end height - max(w, min_unbonding_time)
-		unbondedHeight := activatedDel.EndHeight - minUnbondingTime
+		// the BTC delegation will be unbonded at end height - minUnbondingTime
+		unbondedHeight := activatedDel.EndHeight - h.BTCStakingKeeper.GetParams(h.Ctx).MinUnbondingTimeBlocks
 		events = h.BTCStakingKeeper.GetAllPowerDistUpdateEvents(h.Ctx, unbondedHeight, unbondedHeight)
 		require.Len(t, events, 1)
 		btcDelStateUpdate = events[0].GetBtcDelStateUpdate()
@@ -671,6 +670,7 @@ func TestDoNotGenerateDuplicateEventsAfterHavingCovenantQuorum(t *testing.T) {
 	stakingValue := int64(2 * 10e8)
 	delSK, _, err := datagen.GenRandomBTCKeyPair(r)
 	h.NoError(err)
+	stakingParams := h.BTCStakingKeeper.GetParamsWithVersion(h.Ctx).Params
 	expectedStakingTxHash, msgCreateBTCDel, actualDel, _, _, _, err := h.CreateDelegation(
 		r,
 		delSK,
@@ -681,24 +681,20 @@ func TestDoNotGenerateDuplicateEventsAfterHavingCovenantQuorum(t *testing.T) {
 		0,
 		0,
 		false,
+		false,
 	)
 	h.NoError(err)
 	/*
 		at this point, there should be 1 event that BTC delegation
-		will become expired at end height - w
+		will become expired at end height - min_unbonding_time
 	*/
 	// there exists no event at the current BTC tip
 	btcTip := btclcKeeper.GetTipInfo(h.Ctx)
 	events := h.BTCStakingKeeper.GetAllPowerDistUpdateEvents(h.Ctx, btcTip.Height, btcTip.Height)
 	require.Len(t, events, 0)
 
-	btckptParams := btccKeeper.GetParams(h.Ctx)
-	stakingParams := h.BTCStakingKeeper.GetParamsWithVersion(h.Ctx).Params
-
-	minUnbondingTime := types.MinimumUnbondingTime(&stakingParams, &btckptParams)
-
-	// the BTC delegation will be unbonded at end height - max(w, min_unbonding_time)
-	unbondedHeight := actualDel.EndHeight - minUnbondingTime
+	// the BTC delegation will be unbonded at end height - min_unbonding_time
+	unbondedHeight := actualDel.EndHeight - stakingParams.MinUnbondingTimeBlocks
 	events = h.BTCStakingKeeper.GetAllPowerDistUpdateEvents(h.Ctx, unbondedHeight, unbondedHeight)
 	require.Len(t, events, 1)
 	btcDelStateUpdate := events[0].GetBtcDelStateUpdate()