+++ x/incentive/keeper/reward_gauge.go
@@ -2,6 +2,7 @@ package keeper
 
 import (
 	"context"
+
 	"cosmossdk.io/store/prefix"
 	"github.com/babylonlabs-io/babylon/x/incentive/types"
 	"github.com/cosmos/cosmos-sdk/runtime"
@@ -66,7 +67,7 @@ func (k Keeper) GetRewardGauge(ctx context.Context, sType types.StakeholderType,
 }
 
 // rewardGaugeStore returns the KVStore of the reward gauge of a stakeholder
-// of a given type {submitter, reporter, finality provider, BTC delegation}
+// of a given type {finality provider or BTC delegation}
 // prefix: RewardGaugeKey
 // key: (stakeholder type || stakeholder address)
 // value: reward gauge